requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.2
black==23.12.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
psycopg==3.1.15
psycopg-binary==3.1.15
psycopg-pool==3.2.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
